requirements.txt
arguably
mysql-connector-python
requests
tqdm